+++ lib/src/data/mappers/embed_mapper.dart
@@ -9,11 +9,20 @@ class EmbedMapper implements BlockMapper<EmbedBlock> {
 
   @override
   EmbedBlock fromJson(Map<String, dynamic> data) => EmbedBlock(
-        service: (data['service'] as String?) ?? '',
-        source: (data['source'] as String?) ?? '',
-        embed: (data['embed'] as String?) ?? '',
-        width: data['width'] as int?,
-        height: data['height'] as int?,
-        caption: data['caption'] as String?,
+        service: data['service'] is String ? data['service'] as String : '',
+        source: data['source'] is String ? data['source'] as String : '',
+        embed: data['embed'] is String ? data['embed'] as String : '',
+        width: switch (data['width']) {
+          int v => v,
+          num v => v.toInt(),
+          _ => null,
+        },
+        height: switch (data['height']) {
+          int v => v,
+          num v => v.toInt(),
+          _ => null,
+        },
+        caption:
+            data['caption'] is String ? data['caption'] as String : null,
       );
 }